# incidence-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Plataforma de Gestión de Incidencias
  description: API REST para la gestión de incidencias empresariales.
  version: 1.0.0

servers:
  - url: /api/v1

tags:
  - name: Usuarios
    description: Gestión de usuarios del sistema
  - name: Incidencias
    description: Gestión de incidencias

paths:

  # ── USUARIOS ────────────────────────────────────────────────────────────────

  /usuarios:
    get:
      tags: [Usuarios]
      summary: Listar todos los usuarios
      operationId: listarUsuarios
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioDTO'

    post:
      tags: [Usuarios]
      summary: Crear un nuevo usuario
      operationId: crearUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioCreateDTO'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioDTO'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

  /usuarios/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64

    get:
      tags: [Usuarios]
      summary: Obtener usuario por ID
      operationId: obtenerUsuario
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioDTO'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

    put:
      tags: [Usuarios]
      summary: Actualizar usuario
      operationId: actualizarUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioCreateDTO'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioDTO'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

    delete:
      tags: [Usuarios]
      summary: Eliminar usuario
      operationId: eliminarUsuario
      responses:
        '204':
          description: Usuario eliminado
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

  # ── INCIDENCIAS ──────────────────────────────────────────────────────────────

  /incidencias:
    get:
      tags: [Incidencias]
      summary: Listar todas las incidencias
      operationId: listarIncidencias
      responses:
        '200':
          description: Lista de incidencias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IncidenciaDTO'

    post:
      tags: [Incidencias]
      summary: Crear una nueva incidencia
      operationId: crearIncidencia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidenciaCreateDTO'
      responses:
        '201':
          description: Incidencia creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidenciaDTO'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '404':
          description: Usuario asignado no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

  /incidencias/buscar:
    get:
      tags: [Incidencias]
      summary: Búsqueda avanzada de incidencias con filtros opcionales y paginación
      operationId: buscarIncidencias
      parameters:
        - name: usuarioId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: estado
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/EstadoIncidencia'
        - name: prioridad
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Prioridad'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
            example: fechaCreacion,desc
      responses:
        '200':
          description: Página de incidencias filtradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageIncidenciaDTO'

  /incidencias/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64

    get:
      tags: [Incidencias]
      summary: Obtener incidencia por ID
      operationId: obtenerIncidencia
      responses:
        '200':
          description: Incidencia encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidenciaDTO'
        '404':
          description: Incidencia no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

    put:
      tags: [Incidencias]
      summary: Actualizar incidencia
      operationId: actualizarIncidencia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidenciaCreateDTO'
      responses:
        '200':
          description: Incidencia actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidenciaDTO'
        '404':
          description: Incidencia no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

    delete:
      tags: [Incidencias]
      summary: Eliminar incidencia
      operationId: eliminarIncidencia
      responses:
        '204':
          description: Incidencia eliminada
        '404':
          description: Incidencia no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

  /incidencias/{id}/sugerencia:
    post:
      tags: [Incidencias]
      summary: Generar sugerencia de resolución mediante IA (GitHub Copilot CLI)
      operationId: generarSugerencia
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: Sugerencia generada y persistida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SugerenciaIADTO'
        '404':
          description: Incidencia no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '503':
          description: GitHub Copilot CLI no disponible en el sistema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

# ── SCHEMAS ───────────────────────────────────────────────────────────────────

components:
  schemas:

    Rol:
      type: string
      enum: [ADMIN, SOPORTE, USUARIO]

    EstadoIncidencia:
      type: string
      enum: [ABIERTA, EN_PROGRESO, CERRADA]

    Prioridad:
      type: string
      enum: [BAJA, MEDIA, ALTA, CRITICA]

    UsuarioDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        email:
          type: string
          format: email
        rol:
          $ref: '#/components/schemas/Rol'
        fechaAlta:
          type: string
          format: date

    UsuarioCreateDTO:
      type: object
      required: [nombre, email, rol]
      properties:
        nombre:
          type: string
        email:
          type: string
          format: email
        rol:
          $ref: '#/components/schemas/Rol'

    IncidenciaDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        titulo:
          type: string
        descripcion:
          type: string
        estado:
          $ref: '#/components/schemas/EstadoIncidencia'
        prioridad:
          $ref: '#/components/schemas/Prioridad'
        fechaCreacion:
          type: string
          format: date-time
        fechaActualizacion:
          type: string
          format: date-time
        usuarioAsignado:
          $ref: '#/components/schemas/UsuarioDTO'

    IncidenciaCreateDTO:
      type: object
      required: [titulo, estado, prioridad, usuarioId]
      properties:
        titulo:
          type: string
        descripcion:
          type: string
        estado:
          $ref: '#/components/schemas/EstadoIncidencia'
        prioridad:
          $ref: '#/components/schemas/Prioridad'
        usuarioId:
          type: integer
          format: int64

    SugerenciaIADTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        incidenciaId:
          type: integer
          format: int64
        promptEjecutado:
          type: string
        respuesta:
          type: string
        fechaGeneracion:
          type: string
          format: date-time

    PageIncidenciaDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/IncidenciaDTO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer

    ErrorDTO:
      type: object
      properties:
        status:
          type: integer
        mensaje:
          type: string
        timestamp:
          type: string
          format: date-time